Refactoring examples
