+++ CMakeLists.txt
@@ -40,7 +40,9 @@ set(LIBRARY_OUTPUT_PATH ${CMAKE_BINARY_DIR})
              Tracker/LocalTracker.cpp
              Tracker/Kalman.cpp)
 
-  set(folder_header nms.h
+  set(folder_header MouseExample.h
+             VideoExample.h
+             nms.h
              Detector/Detector.h
              Detector/BackgroundSubtract.h
              Detector/vibe_src/vibe.hpp
